Move semantic indexing tests into backend suite
