+++ services/site/tests/__tests__/semantic-search-cloudflare-vectorize.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, vi } from 'vitest'
-import { vectorizeDeleteByIds } from '../cloudflare.ts'
+import { vectorizeDeleteByIds } from '../../../../other/semantic-search/cloudflare.ts'
 
 const vectorizeConfig = {
 	accountId: 'account-id',
